requirements.txt
# ═══════════════════════════════════════════════════════════════
# DoshaNet v2 — Python Dependencies
#
# PyTorch CPU-only wheel is ~700 MB smaller than the default.
# This is critical for Render free tier (512 MB RAM limit).
# The extra-index-url line tells pip where to find the cpu builds.
# ═══════════════════════════════════════════════════════════════

--extra-index-url https://download.pytorch.org/whl/cpu

# ── Core ML ──────────────────────────────────────────────────
torch==2.1.2+cpu
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.3.2

# ── Image processing ──────────────────────────────────────────
Pillow==10.3.0

# ── Web framework ─────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Explainability ────────────────────────────────────────────
shap==0.45.0

# ── Data ──────────────────────────────────────────────────────
pandas==2.2.2
